Revamp

+++ pc/src/main/java/com/cts/mfrp/pc/dto/DemandAnalyticsResponseDto.java
@@ -11,7 +11,7 @@ public class DemandAnalyticsResponseDto {
     private String medicineId;
     private String medicineName;
     private String genericName;
-    private Integer searchCount;
-    private Integer reservationCount;
-    private LocalDate periodDate;
+    private Long searchCount;        // total search appearances in the window
+    private Long reservationCount;   // total reservations in the window
+    private LocalDate periodDate;    // most recent day this medicine saw activity
 }
